% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/format.prediction.r
\name{format.prediction}
\alias{format.prediction}
\alias{format.prediction.BinaryTree}
\alias{format.prediction.default}
\title{Standardize result of predict method.}
\usage{
format.prediction(object, prediction)

\method{format.prediction}{default}(object, prediction)

\method{format.prediction}{BinaryTree}(object, prediction)
}
\arguments{
\item{object}{model object used for the prediction.}

\item{prediction}{result of \emph{predict} function.}
}
\value{
a vector containing reuslt of \emph{predict} method.
}
\description{
This is a generic function which standardize result of \emph{predict} method.
	This function is an internal function and not intended to be used directly by users.


}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default S3 method.

\item \code{BinaryTree}: BinaryTree class of \code{\link[party]{ctree}} function in \emph{party} package.
}}

