% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modify.args.model.r
\name{modify.args.model}
\alias{modify.args.model}
\title{Supporting generic function which modify}
\usage{
modify.args.model(object, args.model)
}
\arguments{
\item{object}{idol object}
}
\description{
This function calcuate effective bust cap size ratio of your idol
when you can know her cap size.
}
\examples{
my.idol     <- new("Idol")
my.idol.ecr <- effective.capsize(my.idol)
}
\keyword{bust}

