% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predict.gamm.r
\name{predict.gamm}
\alias{predict.gamm}
\title{predict method for gamm object.}
\usage{
\method{predict}{gamm}(object, newdata, type = "link", se.fit = FALSE,
  terms = NULL, block.size = NULL, newdata.guaranteed = FALSE,
  na.action = na.pass, unconditional = FALSE, ...)
}
\description{
predict method for gamm object.
}

